JWT work

+++ src/authorization/claims.rs
@@ -3,6 +3,6 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Claims {
     pub sub: String,
-    pub exp: u64,
+    pub exp: u128,
     pub claims: Vec<String>,
 }
